requirements.txt
flask
anthropic
pillow
